fix: Address second code review feedback
This commit addresses the feedback from the second code review.

Changes:
- Injected `HashcatApiClient` into `MainViewModel`.
- Split `MainScreen.kt` into smaller files for each tab.
- Replaced `Pair` with a `HashModeInfo` data class in `ApiModels.kt`.
- Made the `getAgents` endpoint in `HashtopolisApiClient.kt` configurable.
- Added error handling for hash mode parsing in `MainViewModel.kt`.
- Used the `use` block for `InputStream` in `MainViewModel.kt`.
- Added unique keys to the `LazyColumn` in `HashtopolisTab.kt`.
- Fixed the theming logic in `Theme.kt`.
- Improved the unit tests.
