+++ app/src/main/java/com/example/hashcatgui/ui/theme/Theme.kt
@@ -10,6 +10,7 @@ import androidx.compose.material3.dynamicLightColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalView
@@ -27,6 +28,18 @@ private val DarkColorScheme = darkColorScheme(
     error = Error
 )
 
+private val LightColorScheme = lightColorScheme(
+    primary = Primary,
+    secondary = Secondary,
+    background = Color(0xFFFFFBFE),
+    surface = Color(0xFFFFFBFE),
+    onPrimary = OnPrimary,
+    onSecondary = OnSecondary,
+    onBackground = Color(0xFF1C1B1F),
+    onSurface = Color(0xFF1C1B1F),
+    error = Error
+)
+
 @Composable
 fun HashcatGUITheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
@@ -40,7 +53,7 @@ fun HashcatGUITheme(
             if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
         }
         darkTheme -> DarkColorScheme
-        else -> DarkColorScheme // Force dark theme
+        else -> LightColorScheme
     }
     val view = LocalView.current
     if (!view.isInEditMode) {
